requirements.txt
streamlit==1.28.2
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
PyPDF2==3.0.1
python-dotenv==1.0.0
langchain==0.0.335
langchain-openai==0.0.2
chromadb==0.4.18
openai==1.3.6
Pillow==10.1.0
pytesseract==0.3.10
sentence-transformers==2.2.2
requests==2.31.0
numpy==1.24.3
pandas==2.1.4